Add view for token supply
