+++ data/guaranteedWL.js
@@ -1,258 +1,42 @@
 module.exports = [
-  {tokenId: 193, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 194, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 195, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 196, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 197, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 198, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 199, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 200, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 201, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 202, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 203, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 204, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 205, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 206, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 207, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 208, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 209, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 210, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 211, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 212, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 213, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 214, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 215, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 216, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 217, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 218, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 219, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 220, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 221, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 222, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 223, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 224, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 225, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 226, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 227, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 228, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 229, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 230, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 231, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 232, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 233, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 234, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 235, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 236, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 237, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 238, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 239, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 240, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 241, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 242, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 243, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 244, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 245, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 246, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 247, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 248, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 249, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 250, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 251, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 252, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 253, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 254, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 255, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 256, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 193, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 194, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 195, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 196, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 197, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 198, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 199, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 200, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 201, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 202, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 203, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 204, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 205, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 206, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 207, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 208, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 209, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 210, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 211, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 212, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 213, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 214, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 215, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 216, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 217, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 218, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 219, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 220, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 221, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 222, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 223, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 224, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 225, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 226, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 227, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 228, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 229, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 230, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 231, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 232, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 233, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 234, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 235, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 236, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 237, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 238, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 239, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 240, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 241, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 242, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 243, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 244, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 245, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 246, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 247, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 248, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 249, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 250, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 251, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 252, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 253, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 254, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 255, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 256, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 193, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 194, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 195, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 196, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 197, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 198, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 199, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 200, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 201, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 202, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 203, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 204, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 205, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 206, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 207, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 208, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 209, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 210, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 211, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 212, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 213, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 214, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 215, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 216, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 217, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 218, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 219, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 220, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 221, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 222, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 223, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 224, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 225, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 226, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 227, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 228, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 229, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 230, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 231, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 232, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 233, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 234, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 235, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 236, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 237, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 238, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 239, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 240, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 241, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 242, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 243, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 244, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 245, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 246, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 247, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 248, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 249, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 250, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 251, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 252, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 253, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 254, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 255, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 256, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 193, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 194, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 195, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 196, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 197, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 198, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 199, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 200, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 201, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 202, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 203, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 204, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 205, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 206, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 207, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 208, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 209, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 210, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 211, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 212, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 213, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 214, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 215, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 216, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 217, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 218, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 219, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 220, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 221, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 222, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 223, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 224, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 225, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 226, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 227, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 228, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 229, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 230, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 231, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 232, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 233, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 234, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 235, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 236, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 237, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 238, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 239, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 240, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 241, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 242, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 243, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 244, wallet: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"},
-  {tokenId: 245, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 246, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 247, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 248, wallet: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"},
-  {tokenId: 249, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 250, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 251, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 252, wallet: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"},
-  {tokenId: 253, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 254, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 255, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
-  {tokenId: 256, wallet: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"},
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+  "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
 ];
