+++ scripts/deploy-factory.js
@@ -5,8 +5,6 @@ const ethers = hre.ethers;
 const DeployUtils = require("./lib/DeployUtils");
 let deployUtils;
 
-const rootLeavesAndProofs = require("../data/rootLeavesAndProofs.json");
-
 async function main() {
   deployUtils = new DeployUtils(ethers);
   require("./consoleLogAlert")();
@@ -18,8 +16,8 @@ async function main() {
 
   console.log("Deploying contracts with the account:", deployer.address, "to", network);
 
-  const genesisToken = await deployUtils.deployProxy("BCGenesisToken", "https://api2.byte.city/genesis/");
-  const oracle = await deployUtils.deployProxy("BCOracleToken", "https://api2.byte.city/oracles/");
+  const genesisToken = await deployUtils.deployProxy("BCGenesisToken", "https://meta.byte.city/genesis/");
+  const oracle = await deployUtils.deployProxy("BCOracleToken", "https://meta.byte.city/oracles/");
   const factory = await deployUtils.deployProxy("BCFactory", genesisToken.address, oracle.address);
   await deployUtils.Tx(genesisToken.setFactory(factory.address, true), "Setting factory address");
   await deployUtils.Tx(oracle.setFactory(factory.address, true), "Setting factory address");
