+++ scripts/deploy-root.js
@@ -3,21 +3,16 @@ const hre = require("hardhat");
 const ethers = hre.ethers;
 
 const DeployUtils = require("./lib/DeployUtils");
+const tree0 = require("../data/rootLeavesAndProofs0.json");
+const tree1 = require("../data/rootLeavesAndProofs1.json");
 let deployUtils;
-
-const indexes = [
-  2, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 2, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 1, 0, 2, 3, 0, 0, 0, 2, 3, 0, 1, 0,
-  1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 2, 0, 1, 0, 0, 0, 1, 3, 0, 0, 2, 0, 0, 1, 3, 1, 0, 0, 1, 1, 0, 1, 2, 1, 0, 0, 0, 2, 2, 1,
-  0, 0, 2, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 1, 0, 0, 1, 2, 1, 0, 0, 0, 1, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1,
-  1, 1, 0, 1, 0, 0, 0, 0, 3, 0, 0, 0, 2, 4,
-];
 
 async function main() {
   deployUtils = new DeployUtils(ethers);
   require("./consoleLogAlert")();
 
-  console.error("Remember to update merkle tree and indexes");
-  process.exit();
+  // console.error("Remember to update merkle tree and indexes");
+  // process.exit();
 
   const chainId = await deployUtils.currentChainId();
   let [deployer] = await ethers.getSigners();
@@ -26,24 +21,11 @@ async function main() {
   console.log("Updating root in factory with:", deployer.address, "on", network);
 
   const factory = await deployUtils.attach("BCFactory");
-  await deployUtils.Tx(factory.setRoot(process.env.ROOT), "Setting the root of the merkle tree");
 
-  let encoded = [];
-  let index77;
-  for (let i = 0; i < indexes.length; i++) {
-    if (i % 77 === 0) {
-      index77 = [];
-    }
-    index77.push(indexes[i]);
-    if (index77.length === 77) {
-      encoded.push(await factory.encode(index77));
-    }
-  }
-  if (index77.length > 0) {
-    encoded.push(await factory.encode(index77));
-  }
-  // console.log(encoded.map(e => e.toString()));
-  await deployUtils.Tx(factory.saveRarityIndex(encoded), "Saving rarity index");
+  const tree0 = require("../data/rootLeavesAndProofs0.json");
+  const tree1 = require("../data/rootLeavesAndProofs1.json");
+
+  await deployUtils.Tx(factory.setRoot("0x" + tree0.root, "0x" + tree1.root), "Setting the roots of the merkle trees");
 }
 
 main()
